Provide RGB colors instead of indexes

+++ src/fheroes2/gui/interface_gamearea.h
@@ -80,7 +80,7 @@ namespace Interface
         void BlitOnTile( Surface &, const Sprite &, const Point & ) const;
 
         void UpdateCyclingPalette( int frame );
-        const std::vector<uint8_t> & GetCyclingPalette() const;
+        const std::vector<uint32_t> & GetCyclingPalette() const;
         MapObjectSprite & GetSpriteCache();
 
         void SetUpdateCursor( void );
@@ -110,7 +110,7 @@ namespace Interface
         int borderSizeX;
         int borderSizeY;
 
-        std::vector<uint8_t> _customPalette;
+        std::vector<uint32_t> _customPalette;
         MapObjectSprite _spriteCache;
 
         enum
